Config Class

Represents the complete configuration for the pipeline.

Usage

Config(
  job_id = character(0),
  task_id = character(0),
  min_reference_date = character(0),
  max_reference_date = character(0),
  report_date = character(0),
  production_date = character(0),
  disease = character(0),
  geo_value = character(0),
  geo_type = character(0),
  seed = integer(0),
  horizon = integer(0),
  model = "EpiNow2",
  config_version = character(0),
  quantile_width = c(0.5, 0.95),
  data = Data(),
  priors = list(),
  parameters = Parameters(),
  sampler_opts = list(),
  exclusions = Exclusions()
)

Arguments

job_id

A string specifying the job.

task_id

A string specifying the task.

min_reference_date

A string representing the minimum reference date. Formatted as "YYYY-MM-DD".

max_reference_date

A string representing the maximum reference date. Formatted as "YYYY-MM-DD".

report_date

A string representing the report date. Formatted as "YYYY-MM-DD".

production_date

A string representing the production date. Formatted as "YYYY-MM-DD".

disease

A string specifying the disease being modeled. One of "COVID-19" or "Influenza".

geo_value

An uppercase, two-character string specifying the geographic value, usually a state or "US" for national data.

geo_type

A string specifying the geographic type, usually "state".

seed

An integer for setting the random seed.

horizon

An integer specifying the forecasting horizon.

model

A string specifying the model to be used.

config_version

A numeric value specifying the configuration version.

quantile_width

A vector of numeric values representing the desired quantiles. Passed to tidybayes::median_qi().

data

An instance of Data class containing data configurations.

priors

A list of lists. The first level should contain the key rt with elements mean and sd and the key gp with element alpha_sd.

parameters

An instance of Parameters class containing parameter configurations.

sampler_opts

A list. The Stan sampler options to be passed through EpiNow2. It has required keys: cores, chains, iter_warmup, iter_sampling, max_treedepth, and adapt_delta.

exclusions

An instance of Exclusions class containing exclusion criteria.

See also