Download specified blobs from Blob Storage and save them in a local dir

Note that I think it might be wise to instead specify a blob prefix, list the blobs, and download all the listed blobs. This would let us have some more flexibility with downloading whole remote directories (like delta tables)

Usage

download_from_azure_blob(blob_names, local_dest, container_name)

Arguments

blob_names

A vector of blobs to donwload from container_name

local_dest

The path to the local directory to save the files in

container_name

The Azure Blob Storage container with blob_names

Value

NULL on success