Calculate low case count diagnostic flag

The diagnostic flag is TRUE if either of the last two weeks of the dataset have fewer than an aggregate 10 cases per week. This aggregation excludes the count from confirmed outliers, which have been set to NA in the data.

Usage

low_case_count_diagnostic(df)

Arguments

df

A dataframe as returned by read_data(). The dataframe must include columns such as reference_date (a date vector) and confirm (the number of confirmed cases per day).

Value

A logical value (TRUE or FALSE) indicating whether either of the last two weeks in the dataset had fewer than 10 cases per week.

Details

This function assumes that the df input dataset has been "completed": that any implicit missingness has been made explicit.

See also