Format PMFs for EpiNow2

Opinionated wrappers around EpiNow2::generation_time_opts(), EpiNow2::delay_opts(), or EpiNow2::dist_spec() that formats the generation interval, delay, or right truncation parameters as an object ready for input to EpiNow2.

Usage

format_generation_interval(pmf)

format_delay_interval(pmf)

format_right_truncation(pmf, data)

Arguments

pmf

As returned by read_disease_parameters(). A PMF vector or an NA, if not applying the PMF to the model fit.

data

in the format returned by read_data()

Value

An EpiNow2::*_opts() formatted object or NA with a message

Details

Delays or right truncation are optional and can be skipped by passing pmf = NA.

See also